Completely dettach IO from parallel request execution
Refactor the queue and server so that the handler holds all
responsibility over the job queue and IO
