+++ bin/console
@@ -13,7 +13,7 @@ def new_doc(source)
   RubyLsp::Document.new(source)
 end
 
-@source = T.let(File.read(File.expand_path("../lib/ruby_lsp/handler.rb", __dir__)), String)
+@source = T.let(File.read(File.expand_path("../lib/ruby_lsp/server.rb", __dir__)), String)
 @document = T.let(new_doc(@source), RubyLsp::Document)
 
 IRB.start(__FILE__)
