+++ lib/ruby_lsp/internal.rb
@@ -4,6 +4,11 @@
 require "sorbet-runtime"
 require "syntax_tree"
 require "language_server-protocol"
+require "benchmark"
 
 require "ruby-lsp"
-require "ruby_lsp/handler"
+require "ruby_lsp/utils"
+require "ruby_lsp/server"
+require "ruby_lsp/executor"
+require "ruby_lsp/requests"
+require "ruby_lsp/store"
